fix: Version info on root command, consistent output style of help
