+++ anchor/src/main.rs
@@ -1,6 +1,12 @@
-use clap::Parser;
+use std::sync::LazyLock;
+
+use clap::{
+    Parser,
+    builder::{Styles, styling::AnsiColor},
+};
 use client::{Client, Node, config};
 use environment::Environment;
+use ethereum_hashing::have_sha_extensions;
 use global_config::{GlobalConfig, GlobalFlags};
 use keygen::Keygen;
 use keysplit::Keysplit;
@@ -13,10 +19,54 @@ use tracing::{Level, error, info};
 use tracing_appender::non_blocking::WorkerGuard;
 use tracing_subscriber::{EnvFilter, Layer, fmt, layer::SubscriberExt, util::SubscriberInitExt};
 use types::EthSpecId;
+use version::VERSION;
 
 mod environment;
 
+pub static SHORT_VERSION: LazyLock<String> = LazyLock::new(|| VERSION.replace("Anchor/", ""));
+pub static LONG_VERSION: LazyLock<String> = LazyLock::new(|| {
+    format!(
+        "{}\n\
+         SHA256 hardware acceleration: {}\n\
+         Profile: {}",
+        SHORT_VERSION.as_str(),
+        have_sha_extensions(),
+        build_profile_name(),
+    )
+});
+
+pub const FLAG_HEADER: &str = "Flags";
+
+fn build_profile_name() -> &'static str {
+    // Nice hack from https://stackoverflow.com/questions/73595435/how-to-get-profile-from-cargo-toml-in-build-rs-or-at-runtime
+    // The profile name is always the 3rd last part of the path (with 1 based indexing).
+    // e.g. /code/core/target/cli/build/my-build-info-9f91ba6f99d7a061/out
+    env!("OUT_DIR")
+        .split(std::path::MAIN_SEPARATOR)
+        .nth_back(3)
+        .unwrap_or("unknown")
+}
+
+pub fn get_color_style() -> Styles {
+    Styles::styled()
+        .header(AnsiColor::Yellow.on_default())
+        .usage(AnsiColor::Green.on_default())
+        .literal(AnsiColor::Green.on_default())
+        .placeholder(AnsiColor::Green.on_default())
+}
+
 #[derive(Parser, Clone, Debug)]
+#[clap(
+    name = "anchor",
+    about = "SSV Validator client. Maintained by Sigma Prime.",
+    author = "Sigma Prime <contact@sigmaprime.io>",
+    long_version = LONG_VERSION.as_str(),
+    version = SHORT_VERSION.as_str(),
+    styles = get_color_style(),
+    next_line_help = true,
+    term_width = 80,
+    display_order = 0,
+)]
 struct Cli {
     #[clap(flatten)]
     pub global_flags: GlobalFlags,
